refactor: upgrade all deps
BREAKING CHANGE: only support node >= 4.0.0
