+++ LICENSE.txt
@@ -1,6 +1,7 @@
 This software is licensed under the MIT License.
 
-Copyright (C) 2013 fengmk2 <fengmk2@gmail.com>
+Copyright(c) 2017 node-modules and other contributors.
+Copyright (C) 2013 - 2016 fengmk2 <fengmk2@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
