+++ package.json
@@ -3,25 +3,24 @@
   "version": "0.0.1",
   "description": "BugFree API Node.js client",
   "main": "index.js",
+  "files": [
+    "index.js"
+  ],
   "scripts": {
-    "test": "make test-all",
-    "blanket": { "pattern": "bugfree/lib" },
-    "travis-cov": { "threshold": 100 }
+    "test": "mocha -t 10000 test/*.test.js",
+    "autod": "autod --prefix '^' -w"
   },
   "dependencies": {
-    "eventproxy": ">=0.2.6",
-    "utility": ">=0.1.9",
-    "urllib": ">=0.5.5",
-    "debug": ">=0.7.4"
+    "debug": "^2.6.8",
+    "eventproxy": "~0.3.5",
+    "urllib": "^2.22.0",
+    "utility": "^1.12.0"
   },
   "devDependencies": {
-    "should": "*",
+    "autod": "^2.8.0",
+    "mocha": "*",
     "pedding": "*",
-    "blanket": "*",
-    "travis-cov": "*",
-    "coveralls": "*",
-    "mocha-lcov-reporter": "*",
-    "mocha": "*"
+    "should": "^11.2.1"
   },
   "homepage": "https://github.com/fengmk2/bugfree",
   "repository": {
@@ -29,7 +28,9 @@
     "url": "git://github.com/fengmk2/bugfree.git"
   },
   "keywords": [
-    "bugfree", "bugfree api", "bug"
+    "bugfree",
+    "bugfree api",
+    "bug"
   ],
   "author": "fengmk2 <fengmk2@gmail.com> (http://fengmk2.github.com)",
   "license": "MIT"
